+++ cli/aws_args.go
@@ -1,0 +1,75 @@
+package cli
+
+import (
+	"flag"
+	"fmt"
+	"log"
+	"net/url"
+	"os"
+
+	"github.com/BSick7/aws-signing/config"
+	"github.com/aws/aws-sdk-go-v2/aws/external"
+)
+
+type AwsArgs struct {
+	use       bool
+	endpoint  string
+	service   string
+	dumpCreds bool
+}
+
+func (a *AwsArgs) AddFlags(flags *flag.FlagSet) {
+	flags.BoolVar(&a.use, "aws", false, "use aws request signing")
+	flags.BoolVar(&a.use, "a", false, "use aws request signing (shorthand)")
+
+	flags.StringVar(&a.endpoint, "aws-endpoint", "", "aws endpoint url")
+	flags.StringVar(&a.endpoint, "e", "", "aws endpoint url (shorthand)")
+
+	flags.StringVar(&a.service, "aws-service", "", "aws service")
+	flags.StringVar(&a.service, "s", "", "aws service (shorthand)")
+
+	flags.BoolVar(&a.dumpCreds, "creds", false, "emit creds")
+}
+
+func (AwsArgs) Options() string {
+	return `
+ -a, --aws                    Use AWS Request Signing
+                              Default: false
+                              Env Var: AWS_SIGNING
+
+ -e, --aws-endpoint <url>     AWS Endpoint URL.
+                              Default: http://localhost:9200
+                              Env Var: AWS_ENDPOINT
+
+ -s, --aws-service <service>  AWS Service.
+                              Default: es
+                              Env Var: AWS_SERVICE
+`
+}
+
+func (a AwsArgs) Config() (config.Aws, error) {
+	var endpointUrl *url.URL
+	if a.endpoint != "" {
+		eu, err := url.Parse(a.endpoint)
+		if err != nil {
+			return config.Aws{}, fmt.Errorf("error parsing endpoint url %q: %s", a.endpoint, err)
+		}
+		endpointUrl = eu
+	}
+
+	return config.Aws{
+		Use:         a.use,
+		EndpointUrl: endpointUrl,
+		Service:     a.service,
+	}, nil
+}
+
+func (a AwsArgs) Dump() {
+	if !a.dumpCreds {
+		return
+	}
+	if _, ok := os.LookupEnv("AWS_SIGNING"); a.use || ok {
+		cfg, _ := external.LoadDefaultAWSConfig()
+		log.Println(cfg.Credentials.Retrieve())
+	}
+}