+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0 (Jul 19, 2018)
+
+  * Rebuilding as CLI and library with own signing transport.
+  * Created reverse proxy facility.
+
 ## 0.1.0 (Jun 25, 2018)
 
   * First build.